A pkgapi object

A pkgapi object. This extends (via inheritance) a plumber object, and so only changes to the plumber API are documented here.

Super classes

plumber::plumber -> pkgapi

Methods

Public methods

pkgapi$new()

pkgapi$handle()

pkgapi$request()

pkgapi$clone()

Inherited methods

plumber::hookable$registerHooks()

plumber::plumber$addAssets()

plumber::plumber$addEndpoint()

plumber::plumber$addFilter()

plumber::plumber$addGlobalProcessor()

plumber::plumber$call()

plumber::plumber$filter()

plumber::plumber$mount()

plumber::plumber$onHeaders()

plumber::plumber$onWSOpen()

plumber::plumber$print()

plumber::plumber$registerHook()

plumber::plumber$route()

plumber::plumber$run()

plumber::plumber$serve()

plumber::plumber$set404Handler()

plumber::plumber$setErrorHandler()

plumber::plumber$setSerializer()

plumber::plumber$swaggerFile()

Method new()

Create a pkgapi object

Usage

pkgapi$new(...)

Arguments

...

Parameters passed to plumber

Method handle()

Handle an endpoint

Usage

pkgapi$handle(endpoint)

Arguments

endpoint

A pkgapi_endpoint object representing an endpoint. Unlike plumber, an R function will not work.

Method request()

Send a request to plumber for debugging

Sends a request to plumber so that the API can be easily tested without running the whole API. The interface here will probably change, and may end up using the interface of httr.

Usage

pkgapi$request(method, path, query = NULL, body = NULL, content_type = NULL)

Arguments

method

Name of HTTP method to use (e.g., GET)

path

Path to send the request to

query

Optional query parameters as a named list or character vector.

body

Optional body (only valid with PUT, POST, etc).

content_type

Optional content type (mime) which can be provided alongside body. If not provided it is set to application/octet-stream if body is raw, or application/json otherwise.

Method clone()

The objects of this class are cloneable with this method.

Usage

pkgapi$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.